[MINOR-20]
* added stuff to README.md
* moved CountObjectsInList from ez_list.py to utilities.py
* edited sort_ to return sorted asc/desc lists with indices in ez_list.py
